Add the defconfig
Coming up with it was fun. I had to go and compare the symbol tables between the stock and the compiled image.

The symbol table is located near the end of the file and looks like a massive block of null-terminated strings. You can extract them, replace every \0 into \r\n and diff them
to see what symbols are missing or wrongly present. This allowed to spot a few anomalies with the defconfig quickly.